feat+fix: russian lowercaser & critical fixes

+++ src/indexer/indexer.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include "../logger/logger.h"
 
-Indexer::Indexer(unsigned long bufferSize, IIndexStorage &indStor)
+Indexer::Indexer(unsigned long long bufferSize, IIndexStorage &indStor)
         : maxBufferSizeInBytes(bufferSize), currentSizeInBytes(0), indexStorage(indStor) {
     Logger::info("Indexer", "Indexer module initialized");
 }
